requirements.txt
matplotlib==3.7.5
matplotlib==3.8.4
PyYAML==6.0.1
PyYAML==6.0.2
PyYAML==6.0.2
tokenizers==0.21.0
torch==2.1.2
torch==2.2.2
torch==2.3.0
torch==2.4.0
torch==2.4.1
torch==2.5.0+cu124
torchsummary==1.5.1
torchvision==0.19.1
torchvision==0.16.2
torchvision==0.17.2
torchvision==0.18.1a0
torchvision==0.19.0
torchvision==0.20.0
torchvision==0.20.0+cu124
tqdm==4.67.0
tqdm==4.66.5
transformers==4.46.2
transformers==4.46.0
